fix longs

+++ src/main/java/com/standard/domain/VendaFiltro.java
@@ -27,7 +27,7 @@ public @Data class VendaFiltro implements Serializable {
 
 	}
 
-	public Integer getCodigo() {
+	public Long getCodigo() {
 		return codigo;
 	}
