+++ src/main/java/com/standard/repository/CategoriaRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CategoriaRepository extends JpaRepository<CategoriaEntity, Integer> {
+public interface CategoriaRepository extends JpaRepository<CategoriaEntity, Long> {
 
 }
