+++ src/main/java/com/standard/repository/MarcaRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MarcaRepository extends JpaRepository<MarcaEntity, Integer> {
+public interface MarcaRepository extends JpaRepository<MarcaEntity, Long> {
 
 }
