+++ src/main/java/com/standard/repository/RetiradaRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RetiradaRepository extends JpaRepository<RetiradaEntity, Integer> {
+public interface RetiradaRepository extends JpaRepository<RetiradaEntity, Long> {
 
 }
