+++ src/main/java/com/standard/repository/SubCategoriaRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SubCategoriaRepository extends JpaRepository<SubCategoriaEntity, Integer> {
+public interface SubCategoriaRepository extends JpaRepository<SubCategoriaEntity, Long> {
 
 }
