+++ src/main/java/com/standard/repository/VendaHasItemProdutoRepository.java
@@ -3,6 +3,6 @@ package com.standard.repository;
 import com.standard.entity.VendaHasItemProdutoEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VendaHasItemProdutoRepository extends JpaRepository<VendaHasItemProdutoEntity, Integer> {
+public interface VendaHasItemProdutoRepository extends JpaRepository<VendaHasItemProdutoEntity, Long> {
 
 }
